Bayesian_Matting_Python/requirements.txt
matplotlib==3.7.1
numba==0.56.4
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
scikit_image==0.19.3
scikit_learn==1.2.2
skimage==0.0
